photowatermark/requirements.txt
Pillow>=9.0.0
tkinterdnd2>=0.3.0  # Optional, for drag and drop functionality